+++ src/ft_printf/ft_snprintf_darwin.c
@@ -6,7 +6,7 @@
 /*   By: stanaka2 <stanaka2@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/10 19:48:52 by stanaka2          #+#    #+#             */
-/*   Updated: 2026/05/01 11:03:25 by stanaka2         ###   ########.fr       */
+/*   Updated: 2026/05/18 20:45:38 by stanaka2         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,5 +28,5 @@ int	ft_snprintf(char *str, size_t size, const char *format, ...)
 	va_end(ap);
 	if (ctx.has_error)
 		return (PF_ERROR);
-	return (ctx.count);
+	return ((int)ctx.count);
 }
