+++ src/ft_printf/ft_sprintf.c
@@ -6,7 +6,7 @@
 /*   By: stanaka2 <stanaka2@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/10 19:49:57 by stanaka2          #+#    #+#             */
-/*   Updated: 2026/05/01 19:34:02 by stanaka2         ###   ########.fr       */
+/*   Updated: 2026/05/18 20:45:48 by stanaka2         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,5 +29,5 @@ int	ft_sprintf(char *str, const char *format, ...)
 	va_end(ap);
 	if (ctx.has_error)
 		return (PF_ERROR);
-	return (ctx.count);
+	return ((int)ctx.count);
 }
